ncurses 5.9 - patch 20120630
+ add --enable-interop to sample package scripts (suggested by Juergen
  Pfeifer).
+ update CF_PATH_SYNTAX macro, from mawk changes.
+ modify mk-0th.awk to allow for generating llib-ltic, etc., though
  some work is needed on cproto to work with lib_gen.c to update
  llib-lncurses.
+ remove redundant getenv() cal in database-iterator leftover from
  cleanup in 20120622 changes (report by Sven Joachim).

+++ test/demo_termcap.c
@@ -29,7 +29,7 @@
 /*
  * Author: Thomas E. Dickey
  *
- * $Id: demo_termcap.c,v 1.20 2012/06/22 22:37:05 tom Exp $
+ * $Id: demo_termcap.c,v 1.21 2012/06/30 22:02:56 tom Exp $
  *
  * A simple demo of the termcap interface.
  */
@@ -88,7 +88,7 @@ save_screen(void)
 static char *
 make_dbitem(char *p, char *q)
 {
-    char *result = malloc(strlen(e_opt) + 2 + p - q);
+    char *result = malloc(strlen(e_opt) + 2 + (size_t) (p - q));
     sprintf(result, "%s=%.*s", e_opt, p - q, q);
     return result;
 }
